mv/rename epic folder
