+++ src/stories/epics/removeUser/index.stories.mdx
@@ -9,7 +9,7 @@ import LinkTo from '@storybook/addon-links/react';
 
 ## Overview
 
-The **RemoveUser** component is a
+The **RemoveUser** component is an
 [SLDS Modal](http://design-system-react-components.herokuapp.com/?path=/story/sldsmodal--modal-with-custom-parent-node)
 used to confirm the removal of a user from an **Epic** when that user has a
 **Task** assigned to them.
